Crypter By Neoz11
This is Crypter designer UD

I use XOR Crypt the file.

Instructions:

Open project CrypterNeoz11 and compile work WindowsApplication3

Compiling Stub Work

Open compilation WindowsApplications3 / bin / *. Exe

Select the file to be encrypted and select the generated stub file

Select the number of random key in value 50 important for it to work

Click on Random Key and encrypt the file.

This work is designed for cryptographic study purposes, I am not responsible for its misuse